Add files via upload
Changed the spectral analysis method from the default (periodogram, part of BinaryFileUnpack.py) to a Lomb-Scargle periodogram for deterministic frequencies. This provided the spectra for a small range at a much higher resolution. From this, the method to determine standard deviation was changed to model the frequency peak as a normal distribution peak, then solve for standard deviation accordingly. This reduced the error of measurement.

For neatness, used a helper method (spec_params()) to return base pressure and the physical H parameter determined by the sensor index used and the conduit diameter of the particular experiment.