Add files via upload
Cleaned up the periodic_range function to return better oscillations faster. Before, it was mainly guesswork to find a working oscillation. Now, it is much more streamlined, as explained in the comments.
Now, the y value is found by the same method as in the hot water oscillations - by referencing the calibration dataset.
Made the function to get all the desired parameters broader. It no longer requires the experiments to be sorted by X, and it will recurse through all directories in the provided parent to find proper experiments and extract their parameters.